+++ punchbowl/level2/tests/test_flow.py
@@ -6,7 +6,7 @@ from astropy.wcs import WCS
 from ndcube import NDCube
 from prefect.testing.utilities import prefect_test_harness
 
-from punchbowl.data.io import write_ndcube_to_fits
+from punchbowl.data.punch_io import write_ndcube_to_fits
 from punchbowl.data.tests.test_io import sample_ndcube
 from punchbowl.level2.flow import ORDER, level2_core_flow
